Move model saving and loadin to a new .py file
